requirements.txt
# Core dependencies
langchain>=1.2.0
langchain-community>=0.4.0
langchain-ollama>=1.0.0
langchain-core>=1.2.0
python-dotenv>=1.0.0

# Tools
duckduckgo-search>=8.0.0
ddgs>=0.1.0

# Observability
arize-phoenix>=12.0.0
openinference-instrumentation-langchain>=0.1.0
opentelemetry-api>=1.39.0
opentelemetry-sdk>=1.39.0
opentelemetry-exporter-otlp>=1.39.0
opentelemetry-exporter-otlp-proto-http>=1.39.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
